backend/requirements.txt
fastapi==0.68.1
uvicorn==0.15.0
python-multipart==0.0.5
transformers==4.30.2
torch==2.0.1
Pillow==9.5.0
diffusers==0.17.1
python-dotenv==1.0.0
requests==2.31.0
aiofiles==23.1.0
python-jose==3.3.0 